test: reorganize Surah tests
